feat(mui): make use of lodash & separate view from form
